@@ -13,7 +13,7 @@ import uk.gov.dvla.poc.forms.BicycleLicenceUpdateForm;
 import uk.gov.dvla.poc.model.BicycleLicence;
 import uk.gov.dvla.poc.repository.BicycleLicenceDynamoDBRepository;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.UUID;
